backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
faiss-cpu>=1.13.0
numpy>=2.0.0
PyPDF2>=3.0.1
httpx>=0.28.0
python-dotenv>=1.0.1
pydantic-settings>=2.7.0
python-multipart>=0.0.20
reportlab>=4.0.0
